requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
borb==2.1.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
idna==3.4
ipykernel==6.17.1
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
pdfminer.six==20221105
pdfplumber==0.7.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyMuPDF==1.21.0
pyparsing==3.0.9
PyPDF2==2.11.2
python-barcode==0.14.0
python-dateutil==2.8.2
pyzmq==24.0.1
qrcode==7.3.1
requests==2.28.1
six==1.16.0
stack-data==0.6.2
tornado==6.2
traitlets==5.6.0
typing_extensions==4.4.0
urllib3==1.26.13
Wand==0.6.10
wcwidth==0.2.5
